fix: fix cases of sample with no manifest (#12653)
* fix: fix cases of sample with no manifest

* test: fix ut

* refactor: remove unused file

---------

+++ packages/fx-core/src/component/driver/teamsApp/utils/ManifestUtils.ts
@@ -413,21 +413,23 @@ export class ManifestUtils {
     maxLength = 25
   ): Promise<Result<undefined, FxError>> {
     const manifestPath = this.getTeamsAppManifestPath(projectPath);
-    const manifest = (await fs.readJson(manifestPath)) as TeamsAppManifest;
-    const shortName = manifest.name.short;
-    let hasSuffix = false;
-    let trimmedName = shortName;
-    if (shortName.includes("${{APP_NAME_SUFFIX}}")) {
-      hasSuffix = true;
-      trimmedName = shortName.replace("${{APP_NAME_SUFFIX}}", "");
+    if (fs.pathExistsSync(manifestPath)) {
+      const manifest = (await fs.readJson(manifestPath)) as TeamsAppManifest;
+      const shortName = manifest.name.short;
+      let hasSuffix = false;
+      let trimmedName = shortName;
+      if (shortName.includes("${{APP_NAME_SUFFIX}}")) {
+        hasSuffix = true;
+        trimmedName = shortName.replace("${{APP_NAME_SUFFIX}}", "");
+      }
+      if (trimmedName.length <= maxLength) return ok(undefined);
+      let newShortName = trimmedName.replace(/\s/g, "").slice(0, maxLength);
+      if (hasSuffix) {
+        newShortName += "${{APP_NAME_SUFFIX}}";
+      }
+      manifest.name.short = newShortName;
+      await fs.writeFile(manifestPath, JSON.stringify(manifest, null, 2));
     }
-    if (trimmedName.length <= maxLength) return ok(undefined);
-    let newShortName = trimmedName.replace(/\s/g, "").slice(0, maxLength);
-    if (hasSuffix) {
-      newShortName += "${{APP_NAME_SUFFIX}}";
-    }
-    manifest.name.short = newShortName;
-    await fs.writeFile(manifestPath, JSON.stringify(manifest, null, 2));
     return ok(undefined);
   }
 }
